ref(new-widget-builder-experience): Rename widget folder to widgetBuilder
