+++ static/app/components/dashboards/issueWidgetQueriesForm.tsx
@@ -13,7 +13,7 @@ import {explodeField, generateFieldAsString} from 'sentry/utils/discover/fields'
 import withApi from 'sentry/utils/withApi';
 import withIssueTags from 'sentry/utils/withIssueTags';
 import {DisplayType, WidgetQuery, WidgetType} from 'sentry/views/dashboardsV2/types';
-import {generateIssueWidgetOrderOptions} from 'sentry/views/dashboardsV2/widget/issueWidget/utils';
+import {generateIssueWidgetOrderOptions} from 'sentry/views/dashboardsV2/widgetBuilder/issueWidget/utils';
 import {generateFieldOptions} from 'sentry/views/eventsV2/utils';
 import IssueListSearchBar from 'sentry/views/issueList/searchBar';
 import {IssueSortOptions} from 'sentry/views/issueList/utils';
