+++ static/app/views/dashboardsV2/widgetBuilder/choseDataStep.tsx
