+++ static/app/views/dashboardsV2/widgetBuilder/eventWidget/index.tsx
