+++ static/app/views/dashboardsV2/widgetBuilder/eventWidget/queries.tsx
