+++ static/app/views/dashboardsV2/widgetBuilder/issueWidget/fields.tsx
