+++ static/app/views/dashboardsV2/widgetBuilder/metricWidget/chart.tsx
