+++ static/app/views/dashboardsV2/widgetBuilder/metricWidget/index.tsx
