+++ static/app/views/dashboardsV2/widgetBuilder/metricWidget/queries.tsx
