+++ static/app/views/dashboardsV2/widgetBuilder/metricWidget/searchQueryField.tsx
